FGN_requirements.txt
absl-py==0.12.0
alabaster==0.7.12
astunparse==1.6.3
attrs==20.3.0
Babel==2.9.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cxxfilt==0.2.2
cycler==0.10.0
decorator==4.4.2
docutils==0.16
flake8==3.9.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.7.3
iniconfig==1.1.1
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
olefile
opt-einsum==3.3.0
packaging==20.9
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
tifffile==2021.3.17
toml==0.10.2
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.59.0
typing-extensions
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
